Put all api stuff into detailScreen
